Adds  param to stac_server module
There are times when we don't use a custom domain, and don't need any other features of Cloudfront, that we are fine deploying only an apigw as the entrypoint for stac-server. This is typically false, because in most cases you would have a custom domain. The 'skip' here is intentional (vs setting it to a default false) to imply that you would intentionally be doing something you don't typically do