Add additional security groups to private API GW VPCes (#109)
Add an optional list of security group IDs that'll be applied to the VPC
interface endpoints of a PRIVATE-type cirrus API Gateway. These security
groups are in addition to the security groups that allow traffic from
the private subnet CIDR blocks. Only applicable when the api_rest_type
is PRIVATE.